Updated the CONFIGFILE and CACHEFILE paths
This allows the script to locate the CONFIGFILE and CACHEFILE relative to the scripts location instead of relative to the location of the user when executing the script. Otherwise, executing the script from outside of the script's directory causes the script to not be able to find the two files.